+++ hasktorch/src/Torch/Typed/NN.hs
@@ -29,6 +29,7 @@ import qualified Torch.Autograd                as A
 import qualified Torch.Tensor                  as A
 import qualified Torch.DType                   as D
 import qualified Torch.Device                  as D
+import           Torch.Typed.Aux
 import           Torch.Typed.Factories
 import           Torch.Typed.Native
 import           Torch.Typed.Tensor
